Use exec when launching Gradle
