+++ packages/build-tools/src/android/gradle.ts
@@ -1,12 +1,10 @@
 import { Android, Env, Job, Platform } from '@expo/eas-build-job';
 import { bunyan } from '@expo/logger';
-import spawn, { SpawnPromise, SpawnResult } from '@expo/turtle-spawn';
-import assert from 'assert';
 import fs from 'fs-extra';
 import path from 'path';
 
 import { BuildContext } from '../context';
-import { getParentAndDescendantProcessPidsAsync } from '../utils/processes';
+import { runGradleCommand as runGradleCommandInternal } from '../utils/gradle';
 
 export async function ensureLFLineEndingsInGradlewScript<TJob extends Job>(
   ctx: BuildContext<TJob>
@@ -28,67 +26,16 @@ export async function runGradleCommand(
     extraEnv,
   }: { logger: bunyan; gradleCommand: string; androidDir: string; extraEnv?: Env }
 ): Promise<void> {
-  logger.info(`Running 'gradlew ${gradleCommand}' in ${androidDir}`);
-  await fs.chmod(path.join(androidDir, 'gradlew'), 0o755);
-  const verboseFlag = ctx.env['EAS_VERBOSE'] === '1' ? '--info' : '';
-
-  const spawnPromise = spawn(
-    'bash',
-    ['-c', `./gradlew ${gradleCommand} --profile ${verboseFlag}`],
-    {
-      cwd: androidDir,
-      logger,
-      lineTransformer: (line?: string) => {
-        if (!line || /^\.+$/.exec(line)) {
-          return null;
-        } else {
-          return line;
-        }
-      },
-      env: {
-        ...ctx.env,
-        ...extraEnv,
-        ...resolveVersionOverridesEnvs(ctx),
-        LC_ALL: 'C.UTF-8',
-      },
-    }
-  );
-  if (ctx.env.EAS_BUILD_RUNNER === 'eas-build' && process.platform === 'linux') {
-    adjustOOMScore(spawnPromise, logger);
-  }
-
-  await spawnPromise;
-}
-
-/**
- * OOM Killer sometimes kills worker server while build is exceeding memory limits.
- * `oom_score_adj` is a value between -1000 and 1000 and it defaults to 0.
- * It defines which process is more likely to get killed (higher value more likely).
- *
- * This function sets oom_score_adj for Gradle process and all its child processes.
- */
-function adjustOOMScore(spawnPromise: SpawnPromise<SpawnResult>, logger: bunyan): void {
-  setTimeout(
-    async () => {
-      try {
-        assert(spawnPromise.child.pid);
-        const pids = await getParentAndDescendantProcessPidsAsync(spawnPromise.child.pid);
-        await Promise.all(
-          pids.map(async (pid: number) => {
-            // Value 800 is just a guess here. It's probably higher than most other
-            // process. I didn't want to set it any higher, because I'm not sure if OOM Killer
-            // can start killing processes when there is still enough memory left.
-            const oomScoreOverride = 800;
-            await fs.writeFile(`/proc/${pid}/oom_score_adj`, `${oomScoreOverride}\n`);
-          })
-        );
-      } catch (err: any) {
-        logger.debug({ err, stderr: err?.stderr }, 'Failed to override oom_score_adj');
-      }
+  await runGradleCommandInternal({
+    androidDir,
+    env: {
+      ...ctx.env,
+      ...extraEnv,
+      ...resolveVersionOverridesEnvs(ctx),
     },
-    // Wait 20 seconds to make sure all child processes are started
-    20000
-  );
+    gradleCommand,
+    logger,
+  });
 }
 
 // Version envs should be set at the beginning of the build, but when building
